Organise item and resource type item models
- Relocated the item type and resource type item models
