+++ app/Models/ItemType/SimpleExpense.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Models;
+namespace App\Models\ItemType;
 
 use App\Utilities\General;
 use App\Utilities\Model as ModelUtility;
@@ -17,13 +17,13 @@ use Illuminate\Support\Facades\DB;
  * @copyright G3D Development Limited 2018-2019
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ItemTypeSimpleExpense extends Model
+class SimpleExpense extends Model
 {
     protected $table = 'item_type_simple_expense';
 
     protected $guarded = ['id', 'updated_at', 'created_at'];
 
-    public function instance(int $item_id): ?ItemTypeSimpleExpense
+    public function instance(int $item_id): ?SimpleExpense
     {
         return $this->where('item_id', '=', $item_id)->
             select(
@@ -35,12 +35,12 @@ class ItemTypeSimpleExpense extends Model
     /**
      * Convert the model instance to an array for use with the transformer
      *
-     * @param Item $item
+     * @param $item
      * @param Model $item_type
      *
      * @return array
      */
-    public function instanceToArray(Item $item, Model $item_type): array
+    public function instanceToArray($item, Model $item_type): array
     {
         return [
             'item_id' => $item->id,
