+++ app/Models/ResourceTypeItemType/SimpleExpense.php
@@ -1,10 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Models;
+namespace App\Models\ResourceTypeItemType;
 
 use App\Utilities\General;
-use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use Illuminate\Support\Facades\DB;
@@ -17,11 +16,11 @@ use Illuminate\Support\Facades\DB;
  * @copyright G3D Development Limited 2018-2019
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ResourceTypeItemTypeAllocatedExpense extends Model
+class SimpleExpense extends Model
 {
     protected $table = 'item';
 
-    protected $guarded = ['id', 'actualised_total', 'created_at', 'updated_at'];
+    protected $guarded = ['id', 'created_at', 'updated_at'];
 
     /**
      * Return the total number of items for the requested resource type
@@ -38,19 +37,19 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
         array $search_conditions = []
     ): int
     {
-        $collection = $this->join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+        $collection = $this->join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join('resource', 'item.resource_id', 'resource.id')->
             join('resource_type', 'resource.resource_type_id', 'resource_type.id')->
             where('resource_type.id', '=', $resource_type_id);
 
         if (array_key_exists('year', $parameters_collection) === true &&
             $parameters_collection['year'] !== null) {
-            $collection->where(DB::raw('YEAR(item_type_allocated_expense.effective_date)'), '=', $parameters_collection['year']);
+            $collection->where(DB::raw('YEAR(item_type_simple_expense.effective_date)'), '=', $parameters_collection['year']);
         }
 
         if (array_key_exists('month', $parameters_collection) === true &&
             $parameters_collection['month'] !== null) {
-            $collection->where(DB::raw('MONTH(item_type_allocated_expense.effective_date)'), '=', $parameters_collection['month']);
+            $collection->where(DB::raw('MONTH(item_type_simple_expense.effective_date)'), '=', $parameters_collection['month']);
         }
 
         if (array_key_exists('category', $parameters_collection) === true &&
@@ -70,18 +69,8 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
 
         if (count($search_conditions) > 0) {
             foreach ($search_conditions as $field => $search_term) {
-                $collection->where('item_type_allocated_expense.' . $field, 'LIKE', '%' . $search_term . '%');
+                $collection->where('item_type_simple_expense.' . $field, 'LIKE', '%' . $search_term . '%');
             }
-        }
-
-        if (
-            array_key_exists('include-unpublished', $parameters_collection) === false ||
-            $parameters_collection['include-unpublished'] === false
-        ) {
-            $collection->where(function ($collection) {
-                $collection->whereNull('item_type_allocated_expense.publish_after')->
-                    orWhereRaw('item_type_allocated_expense.publish_after < NOW()');
-            });
         }
 
         return $collection->count();
@@ -114,16 +103,14 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
             'resource.name AS resource_name',
             'resource.description AS resource_description',
             'item.id AS item_id',
-            'item_type_allocated_expense.name AS item_name',
-            'item_type_allocated_expense.description AS item_description',
-            'item_type_allocated_expense.effective_date AS item_effective_date',
-            'item_type_allocated_expense.total AS item_total',
-            'item_type_allocated_expense.percentage AS item_percentage',
-            'item_type_allocated_expense.actualised_total AS item_actualised_total',
+            'item_type_simple_expense.name AS item_name',
+            'item_type_simple_expense.description AS item_description',
+            'item_type_simple_expense.effective_date AS item_effective_date',
+            'item_type_simple_expense.total AS item_total',
             'item.created_at AS item_created_at'
         ];
 
-        $collection = $this->join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+        $collection = $this->join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join('resource', 'item.resource_id', 'resource.id')->
             join('resource_type', 'resource.resource_type_id', 'resource_type.id')->
             where('resource_type.id', '=', $resource_type_id);
@@ -171,12 +158,12 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
 
         if (array_key_exists('year', $parameters_collection) === true &&
             $parameters_collection['year'] !== null) {
-            $collection->where(DB::raw('YEAR(item_type_allocated_expense.effective_date)'), '=', $parameters_collection['year']);
+            $collection->where(DB::raw('YEAR(item_type_simple_expense.effective_date)'), '=', $parameters_collection['year']);
         }
 
         if (array_key_exists('month', $parameters_collection) === true &&
             $parameters_collection['month'] !== null) {
-            $collection->where(DB::raw('MONTH(item_type_allocated_expense.effective_date)'), '=', $parameters_collection['month']);
+            $collection->where(DB::raw('MONTH(item_type_simple_expense.effective_date)'), '=', $parameters_collection['month']);
         }
 
         if (array_key_exists('category', $parameters_collection) === true &&
@@ -199,18 +186,8 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
 
         if (count($search_conditions) > 0) {
             foreach ($search_conditions as $field => $search_term) {
-                $collection->where('item_type_allocated_expense.' . $field, 'LIKE', '%' . $search_term . '%');
+                $collection->where('item_type_simple_expense.' . $field, 'LIKE', '%' . $search_term . '%');
             }
-        }
-
-        if (
-            array_key_exists('include-unpublished', $parameters_collection) === false ||
-            $parameters_collection['include-unpublished'] === false
-        ) {
-            $collection->where(function ($collection) {
-                $collection->whereNull('item_type_allocated_expense.publish_after')->
-                    orWhereRaw('item_type_allocated_expense.publish_after < NOW()');
-            });
         }
 
         if (count($sort_parameters) > 0) {
@@ -219,13 +196,11 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
                     case 'created':
                         $collection->orderBy('item.created_at', $direction);
                         break;
-
-                    case 'actualised_total':
                     case 'description':
                     case 'effective_date':
                     case 'name':
                     case 'total':
-                        $collection->orderBy('item_type_allocated_expense.' . $field, $direction);
+                        $collection->orderBy('item_type_simple_expense.' . $field, $direction);
                         break;
 
                     default:
@@ -234,7 +209,7 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
                 }
             }
         } else {
-            $collection->orderBy('item_type_allocated_expense.effective_date', 'desc');
+            $collection->orderBy('item_type_simple_expense.effective_date', 'desc');
             $collection->orderBy('item.created_at', 'desc');
         }
 
@@ -249,19 +224,16 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      * Return the summary for all items for the resources in the requested resource type
      *
      * @param int $resource_type_id
-     * @param boolean $include_unpublished = false
      *
      * @return array
      */
-    public function summary(int $resource_type_id, bool $include_unpublished): array
+    public function summary(int $resource_type_id): array
     {
-        $collection = $this->selectRaw('sum(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+        $collection = $this->selectRaw('sum(item_type_simple_expense.total) AS total')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join('resource', 'item.resource_id', 'resource.id')->
             join('resource_type', 'resource.resource_type_id', 'resource_type.id')->
             where('resource_type.id', '=', $resource_type_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
 
         return $collection->
             get()->
@@ -273,23 +245,20 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      * type grouped by resource
      *
      * @param int $resource_type_id
-     * @param boolean $include_unpublished
      *
      * @return array
      */
-    public function resourcesSummary(int $resource_type_id, bool $include_unpublished): array
+    public function resourcesSummary(int $resource_type_id): array
     {
         $collection = $this->selectRaw('
                 resource.id AS id, 
                 resource.name AS `name`, 
-                SUM(item_type_allocated_expense.actualised_total) AS total'
+                SUM(item_type_simple_expense.total) AS total'
             )->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join('resource', 'item.resource_id', 'resource.id')->
             join('resource_type', 'resource.resource_type_id', 'resource_type.id')->
             where('resource_type.id', '=', $resource_type_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
 
         return $collection->groupBy('resource.id')->
             orderBy('name')->
@@ -302,22 +271,19 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      * type grouped by year
      *
      * @param int $resource_type_id
-     * @param boolean $include_unpublished
 
      * @return array
      */
-    public function yearsSummary(int $resource_type_id, bool $include_unpublished): array
+    public function yearsSummary(int $resource_type_id): array
     {
         $collection = $this->selectRaw("
-                YEAR(item_type_allocated_expense.effective_date) as year,
-                SUM(item_type_allocated_expense.actualised_total) AS total"
+                YEAR(item_type_simple_expense.effective_date) as year,
+                SUM(item_type_simple_expense.total) AS total"
             )->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
             where("resource_type.id", "=", $resource_type_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
 
         return $collection->groupBy("year")->
             orderBy("year")->
@@ -335,11 +301,11 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      */
     public function maximumEffectiveDateYear(int $resource_type_id): int
     {
-        $result = $this->from('item_type_allocated_expense')->
-            join('item', 'item_type_allocated_expense.item_id', 'item.id')->
+        $result = $this->from('item_type_simple_expense')->
+            join('item', 'item_type_simple_expense.item_id', 'item.id')->
             join('resource', 'item.resource_id', 'resource.id')->
             where('resource.resource_type_id', '=', $resource_type_id)->
-            selectRaw('YEAR(MAX(`item_type_allocated_expense`.`effective_date`)) AS `year_limit`')->
+            selectRaw('YEAR(MAX(`item_type_simple_expense`.`effective_date`)) AS `year_limit`')->
             first();
 
         if ($result === null) {
@@ -359,11 +325,11 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      */
     public function minimumEffectiveDateYear(int $resource_type_id): int
     {
-        $result = $this->from('item_type_allocated_expense')->
-            join('item', 'item_type_allocated_expense.item_id', 'item.id')->
+        $result = $this->from('item_type_simple_expense')->
+            join('item', 'item_type_simple_expense.item_id', 'item.id')->
             join('resource', 'item.resource_id', 'resource.id')->
             where('resource.resource_type_id', '=', $resource_type_id)->
-            selectRaw('YEAR(MIN(`item_type_allocated_expense`.`effective_date`)) AS `year_limit`')->
+            selectRaw('YEAR(MAX(`item_type_simple_expense`.`effective_date`)) AS `year_limit`')->
             first();
 
         if ($result === null) {
@@ -379,23 +345,20 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      *
      * @param integer $resource_type_id
      * @param integer $year
-     * @param boolean $include_unpublished
      *
      * @return array
      */
-    public function monthsSummary(int $resource_type_id, int $year, bool $include_unpublished): array
+    public function monthsSummary(int $resource_type_id, int $year): array
     {
         $collection = $this->selectRaw("
-                MONTH(item_type_allocated_expense.effective_date) as month, 
-                SUM(item_type_allocated_expense.actualised_total) AS total"
+                MONTH(item_type_simple_expense.effective_date) as month, 
+                SUM(item_type_simple_expense.total) AS total"
             )->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
             where("resource_type.id", "=", $resource_type_id)->
-            where(DB::raw('YEAR(item_type_allocated_expense.effective_date)'), '=', $year);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
+            where(DB::raw('YEAR(item_type_simple_expense.effective_date)'), '=', $year);
 
         return $collection->groupBy("month")->
             orderBy("month")->
@@ -410,29 +373,25 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      * @param integer $resource_type_id
      * @param integer $year
      * @param integer $month
-     * @param boolean $include_unpublished
      *
      * @return array
      */
     public function monthSummary(
         int $resource_type_id,
         int $year,
-        int $month,
-        bool $include_unpublished
+        int $month
     ): array
     {
         $collection = $this->selectRaw("
-                MONTH(item_type_allocated_expense.effective_date) as month, 
-                SUM(item_type_allocated_expense.actualised_total) AS total"
+                MONTH(item_type_simple_expense.effective_date) as month, 
+                SUM(item_type_simple_expense.total) AS total"
             )->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
             where("resource_type.id", "=", $resource_type_id)->
-            where(DB::raw('YEAR(item_type_allocated_expense.effective_date)'), '=', $year)->
-            where(DB::raw('MONTH(item_type_allocated_expense.effective_date)'), '=', $month);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
+            where(DB::raw('YEAR(item_type_simple_expense.effective_date)'), '=', $year)->
+            where(DB::raw('MONTH(item_type_simple_expense.effective_date)'), '=', $month);
 
         return $collection->groupBy("month")->
             orderBy("month")->
@@ -446,23 +405,20 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      *
      * @param integer $resource_type_id
      * @param integer $year
-     * @param boolean $include_unpublished
      *
      * @return array
      */
-    public function yearSummary(int $resource_type_id, int $year, bool $include_unpublished): array
+    public function yearSummary(int $resource_type_id, int $year): array
     {
         $collection = $this->selectRaw("
-                YEAR(item_type_allocated_expense.effective_date) as year, 
-                SUM(item_type_allocated_expense.actualised_total) AS total"
+                YEAR(item_type_simple_expense.effective_date) as year, 
+                SUM(item_type_simple_expense.total) AS total"
             )->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
             where("resource_type.id", "=", $resource_type_id)->
-            where(DB::raw('YEAR(item_type_allocated_expense.effective_date)'), '=', $year);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
+            where(DB::raw('YEAR(item_type_simple_expense.effective_date)'), '=', $year);
 
         return $collection->groupBy("year")->
             orderBy("year")->
@@ -475,29 +431,25 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      * type grouped by category
      *
      * @param integer $resource_type_id
-     * @param boolean $include_unpublished
      *
      * @return array
      */
     public function categoriesSummary(
-        int $resource_type_id,
-        bool $include_unpublished = false
+        int $resource_type_id
     ): array
     {
         $collection = $this->selectRaw('
                 category.id, 
                 category.name AS name, 
                 category.description AS description,
-                SUM(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+                SUM(item_type_simple_expense.total) AS total')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
             join("item_category", "item_category.item_id", "item.id")->
             join("category", "category.id", "item_category.category_id")->
             where("category.resource_type_id", "=", $resource_type_id)->
             where("resource_type.id", "=", $resource_type_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
 
         return $collection->groupBy("category.id")->
             orderBy("name")->
@@ -511,22 +463,20 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      *
      * @param integer $resource_type_id
      * @param integer $category_id
-     * @param boolean $include_unpublished
      *
      * @return array
      */
     public function categorySummary(
         int $resource_type_id,
-        int $category_id,
-        bool $include_unpublished = false
+        int $category_id
     ): array
     {
         $collection = $this->selectRaw('
                 category.id, 
                 category.name AS name, 
                 category.description, 
-                SUM(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+                SUM(item_type_simple_expense.total) AS total')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
             join("item_category", "item_category.item_id", "item.id")->
@@ -534,8 +484,6 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
             where("category.resource_type_id", "=", $resource_type_id)->
             where("resource_type.id", "=", $resource_type_id)->
             where("category.id", '=', $category_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
 
         return $collection->groupBy("category.id")->
             orderBy("name")->
@@ -549,13 +497,12 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
         int $subcategory_id = null,
         int $year = null,
         int $month = null,
-        array $search_parameters = [],
-        bool $include_unpublished = false
+        array $search_parameters = []
     ): array
     {
         $collection = $this->
-            selectRaw('SUM(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+            selectRaw('SUM(item_type_simple_expense.total) AS total')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
             join("item_category", "item_category.item_id", "item.id")->
@@ -571,18 +518,16 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
             $collection->where("sub_category.id", "=", $subcategory_id);
         }
         if ($year !== null) {
-            $collection->whereRaw(DB::raw("YEAR(item_type_allocated_expense.effective_date) = {$year}"));
+            $collection->whereRaw(DB::raw("YEAR(item_type_simple_expense.effective_date) = {$year}"));
         }
         if ($month !== null) {
-            $collection->whereRaw(DB::raw("MONTH(item_type_allocated_expense.effective_date) = {$month}"));
+            $collection->whereRaw(DB::raw("MONTH(item_type_simple_expense.effective_date) = {$month}"));
         }
         if (count($search_parameters) > 0) {
             foreach ($search_parameters as $field => $search_term) {
-                $collection->where('item_type_allocated_expense.' . $field, 'LIKE', '%' . $search_term . '%');
+                $collection->where('item_type_simple_expense.' . $field, 'LIKE', '%' . $search_term . '%');
             }
         }
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
 
         return $collection->get()->
             toArray();
@@ -594,22 +539,20 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      *
      * @param int $resource_type_id
      * @param int $category_id
-     * @param boolean $include_unpublished
      *
      * @return array
      */
     public function subcategoriesSummary(
         int $resource_type_id,
-        int $category_id,
-        bool $include_unpublished = false
+        int $category_id
     ): array
     {
         $collection = $this->selectRaw('
                 sub_category.id, 
                 sub_category.name AS name, 
                 sub_category.description AS description, 
-                SUM(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+                SUM(item_type_simple_expense.total) AS total')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
             join("item_category", "item_category.item_id", "item.id")->
@@ -619,8 +562,6 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
             where("category.resource_type_id", "=", $resource_type_id)->
             where("resource_type.id", "=", $resource_type_id)->
             where("category.id", "=", $category_id);
-
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
 
         return $collection->groupBy("sub_category.id")->
             orderBy("name")->
@@ -635,23 +576,21 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
      * @param int $resource_type_id
      * @param int $category_id
      * @param int $subcategory_id
-     * @param boolean $include_unpublished
      *
      * @return array
      */
     public function subcategorySummary(
         int $resource_type_id,
         int $category_id,
-        int $subcategory_id,
-        bool $include_unpublished = false
+        int $subcategory_id
     ): array
     {
         $collection = $this->selectRaw('
                 sub_category.id, 
                 sub_category.name AS name, 
                 sub_category.description AS description, 
-                SUM(item_type_allocated_expense.actualised_total) AS total')->
-            join('item_type_allocated_expense', 'item.id', 'item_type_allocated_expense.item_id')->
+                SUM(item_type_simple_expense.total) AS total')->
+            join('item_type_simple_expense', 'item.id', 'item_type_simple_expense.item_id')->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
             join("item_category", "item_category.item_id", "item.id")->
@@ -663,31 +602,9 @@ class ResourceTypeItemTypeAllocatedExpense extends Model
             where("category.id", "=", $category_id)->
             where('sub_category.id', '=', $subcategory_id);
 
-        $collection = $this->includeUnpublished($collection, $include_unpublished);
-
         return $collection->groupBy("sub_category.id")->
             orderBy("name")->
             get()->
             toArray();
-    }
-
-    /**
-     * Work out if we should be hiding unpublished items, by default we don't show them
-     *
-     * @param $collection
-     * @param boolean $include_unpublished
-     *
-     * @return Builder
-     */
-    private function includeUnpublished($collection, bool $include_unpublished): Builder
-    {
-        if ($include_unpublished === false) {
-            $collection->where(function ($sql) {
-                $sql->whereNull('item_type_allocated_expense.publish_after')->
-                    orWhereRaw('item_type_allocated_expense.publish_after < NOW()');
-            });
-        }
-
-        return $collection;
     }
 }
